Update patch set 2
Patch Set 2:

Marcus - thanks for pointing out that spacing.  I updated my Eclipse setup to use 4 spaces instead of tabs for indentation and reformatted the pom.xml, so should be good now.

On the other comment ... I was thinking that, when we roll CCSDK from Boron to Carbon, we'd bump up the minor version, since that change isn't backward compatible ... so we'd go from 0.0.1-SNAPSHOT to 0.1.1-SNAPSHOT.  I forgot to change the versions in the carbon pom - just submitted a change to fix that (boron was using range up to 0.0.1, Carbon uses range [0.1, 0.2), which I think should match 0.1.* ... you'd think just 0.1 would do that, but nope ... I tried changing boron to 0.0 as a test and it broke my local build

Patch-set: 2
Label: Code-Review=0