Break dependency loop
Removed import of CCSDK repo artifact BOMs from parent poms
to break a dependency loop.  Including those meant that, when
rolling to a new release version, we'd have to touch parent twice:
first, with the 'old' version of CCDSK repos (since they're not built
yet), and then after they're built, we'd have to update parent
poms to import the rolled-version of the repo BOMs.  And after
that, we'd need to re-merge all the repos, so that they are compiled
against the correct version of sli core.

Change-Id: Id10c29e3f321defb1b911de894cf045693ef90cc
Issue-ID: CCSDK-939
Signed-off-by: Timoney, Dan (dt5972) <dtimoney@att.com>