Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > There seems to be one from Kevin, https://gerrit.onap.org/r/c/logging-analytics/+/102099
> > 
> > Kevin can you confirm if the fix is expected to be included in Frankfurt release ?
> 
> I'll just update the version.  This needs to move forward.

Strike that last comment.  That version change should be done in a separate change to the master branch so that it is not lost. If we truly need it in Frankfurt, we can cherry pick it, but ONLY if it is to
correct a High or Highest severity defect tracked by ONAP integration - no other changes are permitted at this late stage.

The scope of this change is merely a version roll of our code. No other changes should be included - since that version roll is branch specific.

Patch-set: 1
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>