Updates to support use of oparent in non-ODL dependent code
Added missing dependencies needed in oparent and dependencies-bom
to allow oparent pom to be used in ccsdk/sli for code not
directly dependent on OpenDaylight.  Also, updated to latest
springboot release and to latest released version of ONAP oparent
pom.  Note: this change does NOT address changes to compile jars as
java 11.  That change will come with the changes for ODL Aluminum.

Change-Id: I599bae0c32f5ced7eada956a0ed3b57f40461cd5
Issue-ID: CCSDK-3001
Signed-off-by: Dan Timoney <dtimoney@att.com>