Update patch set 6
Patch Set 6: Code-Review-2

I'm doing some local testing with this change - there are changes needed in ccsdk/sli for Aluminum that will need to be coordinated with this merge.  Marking this as -2 for now so we remember to hold this merge until the associated changes in other repos are ready.

Patch-set: 6
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-2