Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I would recommond using https://gerrit.onap.org/r/gitweb?p=ccsdk/parent.git;a=blob;f=springboot/spring-boot-setup/src/main/resources/pom-template.xml;hb=HEAD to generate springboot25 and springboot26 both. see if you can refactor same !

Hi Kapil
I spent many hours trying to get that to template to work. 
I don't think the 25 version uses that template approach either. 
In the end it was Much easier to just re-write a new POM.
The suggestion from Dan Timoney was just to copy the 25 version and edit it. 

BTW, why have we so many parent POMs?   (1, 2, 25 & now 26)
Doesn't this somewhat defeat the idea of Parent POMs?

Patch-set: 1